federated_pneumonia_detection/src/control/federated_new_version/pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "core"
version = "1.0.0"
description = "Federated Learning with PyTorch Lightning and Flower (Quickstart Example)"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.22.0",
    "flwr-datasets[vision]>=0.5.0",
    "pytorch-lightning==2.5.5",
    "torch==2.8.0",
    "torchvision==0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = [
    ".",
]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "federated_pneumonia_detection.src.control.federated_new_version.core.server_app:app"
clientapp = "federated_pneumonia_detection.src.control.federated_new_version.core.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
max-epochs = 2

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation.options]
num-supernodes = 2
